Use HostConfig.Mounts to bind paths into container
When using the binds api the docker engine will auto-create the host
path as a directory if it doesn't exist. This can lead to unintended
side-effects when the source is a file (like when bind-mounting the
docker socket), as seen in the linked issue.

This also handles tmpfs and named volumes previously set without Mounts,
we take Config.Volumes, HostConfig.Binds and HostConfig.Tmpfs and
deduplicate over the full set.

Closes https://github.com/balena-os/balena-supervisor/issues/1231

Change-type: patch
Signed-off-by: Robert Günzler <robertg@balena.io>